requirements.txt
faiss_cpu==1.8.0
huggingface_hub==0.20.2
langchain==0.2.10
langchain_community==0.2.9
langchain_core==0.2.22
matplotlib==3.7.1
scikit_learn==1.4.0
streamlit==1.34.0
transformers==4.36.2
utils==1.0.2